% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark.R
\name{BCHMK}
\alias{BCHMK}
\title{Simulation of a Benchmark data}
\usage{
\code{BCHMK(n=200,p=100,nondes=50,sigmaondes=0.05,sigmay=0.5)}
}
\arguments{
\item{n}{a positive integer. \code{n} is the number of observations.}

\item{p}{a positive integer. \code{p} is the number of variables.}

\item{nondes}{a positive integer. \code{p} is the number of Guassians in the mixture.}

\item{sigmaondes}{the standard deviation of the Gaussians.}

\item{sigmay}{the uncertainty on \code{y}.}

\item{X}{the preditors matrix.}

\item{y}{the response vector.}

\item{y0}{the reponse vector without noise \code{sigmay}.}
}
\description{
The function \code{BCHMK} simulates a mixture of \code{nondes} Gaussians representing a predictors matrix \code{X}
of \code{n} observations and \code{p} variables and a response vector \code{y}.
}
\details{
The predictors matrix \code{X} represents a mixture of \code{nondes} Gaussians characterized by means and
amplitudes chosen randomly and a standard deviation set to \code{sigmaondes}.

The response vector \code{y} is the sum of 4 intervals of \code{X} to which we add noise of \code{sigmay}.
#' @return A \code{list} of the following attributes
}
\examples{
### load dual.spls library
library(dual.spls)
### parameters
n <- 100
p <- 50
nondes <- 20
sigmaondes <- 0.5
data.benchmark=BCHMK(n=n,p=p,nondes=nondes,sigmaondes=sigmaondes)

X <- data.benchmark$X
y <- data.benchmark$y

###plotting the data
plot(X[1,],type='l',ylim=c(0,max(X)),main='Benchmark data', ylab='X',col=1)
for (i in 2:n){
 lines(X[i,],col=i)
}
}
\seealso{
\code{browseVignettes("dual.spls")}
}
\author{
François Wahl Louna Alsouki
}
