% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.spls.calval.R
\name{d.spls.calval}
\alias{d.spls.calval}
\title{Splits data into calibration and validation sets according to wich group belongs each observation}
\usage{
d.spls.calval(X, Xtype, Listecal)
}
\arguments{
\item{X}{a numeric matrix.}

\item{Xtype}{a vector of index specifying to wich group belongs each observation.}

\item{Listecal}{a vector specifying how many observations from each group should be selected as calibration.}
}
\value{
a numeric vector giving the row indices of the input data selected for calibration
}
\description{
The function \code{d.spls.calval} divides the data \code{X} into a calibration and a validation set using
the Kennard and Stone startegy for each group at a time and according to the number of calibration desired
from each group.
}
\seealso{
\code{\link[=d.spls.type]{d.spls.type()}},\code{\link[=d.spls.modifiedKS]{d.spls.modifiedKS()}},\code{browseVignettes("dual.spls")}
}
\author{
Louna Alsouki François Wahl
}
\keyword{internal}
