% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.spls.GLB.R
\name{d.spls.GLB}
\alias{d.spls.GLB}
\title{Dual Sparse Partial Least Squares (Dual-SPLS) regression for the norm
\eqn{\Omega(w)=\sum_{g} \alpha_g \|w \|_2+\sum_{g} \lambda_g \|w_g \|_1} for
\eqn{\sum_{g} \alpha_g=1; \Omega(w_g)=\gamma_g ;\sum_{g} \gamma_g=1}}
\usage{
d.spls.GLB(X, y, ncp, ppnu, indG, gamma, verbose = FALSE)
}
\arguments{
\item{X}{a numeric matrix of predictors values. Each row represents an observation and each column a predictor variable.}

\item{y}{a numeric vector or a one column matrix of responses. It represents the response variable for each converstation.}

\item{ncp}{a positive integer. \code{ncp} is the number of Dual-SPLS components.}

\item{ppnu}{a positive real value or a vector of length the number of groups, in \code{[0,1]}.
\code{ppnu} is the desired proportion of variables to shrink to zero for each component and for each group.}

\item{indG}{a numeric vector of group index for each observation.}

\item{gamma}{a numeric vector of the norm \eqn{\Omega} of each \eqn{w_g} verifying sum(gamma)=1.}

\item{verbose}{a boolean value indicating whether or not to diplay the iterations steps.}
}
\value{
A \code{list} of the following attributes
\item{Xmean}{the mean vector of the predictors matrix \code{X}.}
\item{scores}{the matrix of dimension \eqn{n x ncp} where \code{n} is the number of observations.The \code{scores} represents
the observations in the new component basis computed by the compression step
of the Dual-SPLS.}
\item{loadings}{the matrix of dimension \eqn{p x ncp} that represents the Dual-SPLS components.}
\item{Bhat}{the matrix of dimension \eqn{p x ncp} that regroups the regression coefficients for each component.}
\item{intercept}{the vector of length \eqn{ncp} representing the intercept values for each component.}
\item{fitted.values}{the matrix of dimension \eqn{n x ncp} that represents the predicted values of \code{y}}
\item{residuals}{the matrix of dimension \eqn{n x ncp} that represents the residuals corresponding
to the difference between the responses and the fitted values.}
\item{lambda}{the matrix of dimension \eqn{G x ncp} collecting the parameters of sparsity \eqn{\lambda_g} used to fit the model at each iteration and for each group, where
\eqn{G} is the number of groups.}
\item{alpha}{the matrix of dimension \eqn{G x ncp} collecting the constraint parameters \eqn{\alpha_g}  used to fit the model at each iteration and for each group.}
\item{zerovar}{the matrix of dimension \eqn{G x ncp} representing the number of variables shrinked to zero per component and per group.}
}
\description{
The function \code{d.spls.GLB} performs dimentional reduction as in PLS methodology combined to variable selection using the
Dual-SPLS algorithm with the norm \eqn{\Omega(w)=\sum_{g} \alpha_g \|w \|_2+\sum_{g} \lambda_g \|w_g \|_1} for combined data where
\eqn{\sum_{g} \alpha_g=1; \Omega(w_g)=\gamma_g ;\sum_{g} \gamma_g=1}.
}
\seealso{
\code{\link[=d.spls.GLA]{d.spls.GLA()}},\code{\link[=d.spls.GLC]{d.spls.GLC()}},\code{\link[=d.spls.GL]{d.spls.GL()}},\code{browseVignettes("dual.spls")}
}
\author{
Louna Alsouki François Wahl
}
\keyword{internal}
