% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.spls.listecal.R
\name{d.spls.listecal}
\alias{d.spls.listecal}
\title{Determine the number of observations to select from each cell}
\usage{
d.spls.listecal(Xtype, pcal)
}
\arguments{
\item{Xtype}{a vector of index specifying to which group belongs each observation.}

\item{pcal}{a positive integer between 0 and 100. \code{pcal} is the percentage
of calibration samples to be selected.}
}
\value{
a numeric vector specifying how many observations from each group should be selected as calibration.
}
\description{
The function \code{d.spls.listecal} detemines the number of observations to select as calibration from
each cell.
}
\seealso{
\link{d.spls.type},\link{d.spls.calval},\link{d.spls.split}
}
\author{
Louna Alsouki François Wahl
}
\keyword{internal}
