% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.spls.norm.R
\name{d.spls.norm1}
\alias{d.spls.norm1}
\title{Computing the l1 and l2 norm}
\usage{
d.spls.norm1(x)

d.spls.norm2(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
A positive integer of the norm value.
}
\description{
The internal functions \code{norm1} and \code{norm2} computes the l1 and l2 norm of any vector
}
\author{
Louna Alsouki François Wahl
}
\keyword{internal}
