% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.spls.split.R
\name{d.spls.split}
\alias{d.spls.split}
\title{Splits data into calibration and validation sets according to wich group belongs each observation}
\usage{
d.spls.split(X, Xtype, Listecal = NULL)
}
\arguments{
\item{X}{a numeric matrix.}

\item{Xtype}{a vector of index specifying to which group belongs each observation.}

\item{Listecal}{a vector specifying how many observations from each group should be selected as calibration.}
}
\value{
a numeric vector giving the row indices of the input data selected for calibration
}
\description{
The function \code{d.spls.split} divides the data \code{X} into a calibration and a validation set using
the Kennard and Stone strategy for each group at a time and according to the number of calibration desired
from each group.
}
\seealso{
\link{d.spls.type},\link{d.spls.calval}
}
\author{
Louna Alsouki François Wahl
}
\keyword{internal}
