% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual.spls-package.R
\docType{package}
\name{dual.spls-package}
\alias{dual.spls}
\alias{dual.spls-package}
\title{dual.spls package}
\description{
Provides a series of functions that compute latent sparse components that are used in a regression model.
The optimization problem of the Dual-SPLS regression depends on the norm \eqn{\Omega(w)} chosen.
This procedure computes latent sparse components that are used in a regression model.
#' Indeed, we are searching for \code{w} that goes with
\deqn{\Omega^*(z)=max_w(z^Tw) \text{ s.t. } \Omega(w)=1}
It also suggest a calibration and validation method based on a modified version of the Kennard
and Stone Algorithm and a function that simulates data composed of Guassians mixtures.
}
\seealso{
\code{\link[=d.spls.lasso]{d.spls.lasso()}}, \code{\link[=d.spls.LS]{d.spls.LS()}}, \code{\link[=d.spls.ridge]{d.spls.ridge()}},
\code{\link[=d.spls.GL]{d.spls.GL()}}, \code{browseVignettes("dual.spls")}
}
\author{
Louna Alsouki
}
