% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified.KS.R
\name{modified.KS}
\alias{modified.KS}
\title{Splits data into calibration and validation sets according to wich group belongs each observation}
\usage{
modified.KS(X, Xtype, Listecal)
}
\arguments{
\item{X}{a numeric matrix.}

\item{Xtype}{a vector of index specifying to wich group belongs each observation.}

\item{Listecal}{a vector specifying how many observations from each group should be selected as calibration.}
}
\value{
a numeric vector giving the row indices of the input data selected for calibration
}
\description{
The function \code{modified.KS} divides the data \code{X} into a calibration and a validation set using
the Kennard and Stone startegy for each group at a time and according to the number of calibration desired
from each group
}
\seealso{
\code{\link[=type]{type()}},\code{\link[=calval]{calval()}},\code{browseVignettes("dual.spls")}
}
\keyword{internal}
