% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm1}
\alias{norm1}
\title{Computing the l1 and l2 norm}
\usage{
norm1(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
A positive integer of the norm value.
}
\description{
The internal functions \code{norm1} and \code{norm2} computes the l1 and l2 norm of any vector
}
\seealso{
\code{browseVignettes("dual.spls")}
}
\author{
François Wahl Louna Alsouki
}
\keyword{internal}
