% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.spls.lasso.R
\name{d.spls.lasso}
\alias{d.spls.lasso}
\title{Dual Sparse Partial Least Squares (Dual-SPLS) regression for the norm \eqn{\Omega(w)=\lambda \|w\|_1 + \|w\|_2}}
\usage{
\code{d.spls.lasso(X,y,ncp,pctnu,verbose=FALSE) }
}
\arguments{
\item{X}{a numeric matrix of predictors values. Each row represents an observation and each column a predictor variable.}

\item{y}{a numeric vector or a one column matrix of responses. It represents the response variable for each converstation.}

\item{ncp}{a positive integer. \code{ncp} is the number of Dual-SPLS components.}

\item{pctnu}{a positive real value, in \code{[0,1]}. \code{pctnu} is the desired
proportion of variables to shrink to zero for each component.}

\item{verbose}{a boolean value indicating whether or not to diplay the iterations steps.}

\item{Xmean}{the mean vector of the predictors matrix \code{X}.}

\item{scores}{a matrix of \code{ncp} columns representing the Dual-SPLS components and the same number of rows
as \code{X} representing the observations in the new component basis computed by the compression step
of the Dual-SPLS.}

\item{loadings}{the loadings matrix.}

\item{Bhat}{the regression coefficients matrix for each component.}

\item{intercept}{the intercept value for each component.}

\item{fitted.values}{the matrix of predicted values of \code{y}}

\item{residuals}{the matrix of residuals corresponding to the difference between the response and the fitted values.}

\item{lambda}{the vector of the sparse hyper-parameter used to fit the model at each iteration.}

\item{zerovar}{the vector of the number of variables shrinked to zero per component.}
}
\value{
A \code{list} of the following attributes
}
\description{
The function \code{d.spls.lasso} performs dimentional reduction combined to variable selection using the
Dual-SPLS algorithm with the norm \eqn{\Omega(w)=\lambda \|w\|_1 + \|w\|_2}.
}
\details{
This procdure computes latent sparse components that are used in a regression model.
The optimization problem of the Dual-SPLS regression for the norm \eqn{\Omega(w)=\lambda \|w\|_1 + \|w\|_2} is
similar to the Dual norm defintion of the norm \eqn{\Omega(w)}. Indeed, we are searching for \code{w} that goes
with
\deqn{\Omega^*(z)=max_w(z^Tw) \text{ s.t. } \Omega(w)=1}
The solution of this problem is
\deqn{\dfrac{w_j}{\|w\|_2}=\dfrac{1}{\mu} \delta_j (|z_j|-\nu)_+}
Where
\itemize{
\item \eqn{\delta_j} is the vector of signs of \code{z} and \code{w}
\item \eqn{\mu} is a parameter that guarentees the constraint of \eqn{\Omega(w)=1}
\item \eqn{\nu} is the shrinkage parameter.
}

The parameter \eqn{\nu} is computed adaptively according to the proportion of zero variables desired \code{pctnu}.
}
\examples{
### load dual.spls library
library(dual.spls)
### parameters
n <- 100
p <- 50
nondes <- 20
sigmaondes <- 0.5
data.benchmark=BCHMK(n=n,p=p,nondes=nondes,sigmaondes=sigmaondes)

X <- data.benchmark$X
y <- data.benchmark$y

#fitting the model
mod.dspls <- d.spls.lasso(X=X,y=y,ncp=10,pctnu=0.9,verbose=TRUE)

str(mod.dspls)

### plotting the observed values VS predicted values
plot(y,model.dspls$fitted.values, xlab="Observed values", ylab="Predicted values", main="Observed VS Predicted")
points(-1000:1000,-1000:1000,type='l')

### plotting the regression coefficients
par(mfrow=c(3,1))

nz=Mod.dspls$zerovar[i]
plot(1:dim(X)[2],mod.dspls$Bhat[,i],type='l',
    main=paste(" Dual-SPLS (lasso), ncp =", i, " #0coef =", nz, "/", dim(X)[2]),
    ylab='',xlab='' )
inonz=which(mod.dspls$Bhat[,i]!=0)
points(inonz,mod.dspls$Bhat[inonz,i],col='red',pch=19,cex=0.5)
legend("topright", legend ="non null values", bty = "n", cex = 0.8, col = "red",pch=19)
}
\seealso{
\code{browseVignettes("dual.spls")}
}
\author{
François Wahl Louna Alsouki
}
