% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.spls.NIR.R
\docType{data}
\name{d.spls.NIR}
\alias{d.spls.NIR}
\title{Dual Sparse Partial Least Squares (Dual-SPLS) Near Infrared data}
\format{
A \code{data.frame} of 208 observations. The NIR data contains 1557 variables and the density is a one dimensional vector.
}
\usage{
data(d.spls.NIR)
}
\description{
This dataset contains near infra red spectra of refined petroleum samples, associated with the values of their density.
}
\details{
The NIR data is composed of 1557 variables that represent a regular sequence of the absorbance interval.

This data is useful when building a Dual-SPLS regression with \code{NIR} data as the predictor and \code{density} as a response.
}
\keyword{datasets}
