% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.spls.ridge.R
\name{d.spls.ridge}
\alias{d.spls.ridge}
\title{Dual Sparse Partial Least Squares (Dual-SPLS) regression for the norm \eqn{\Omega(w)=\lambda_1 \|w\|_1 +\lambda_2 \|Xw\|_2 + \|w\|_2}}
\usage{
d.spls.ridge(X,y,ncp,ppnu,nu2,verbose=FALSE)
}
\arguments{
\item{X}{a numeric matrix of predictors values. Each row represents an observation and each column a predictor variable.}

\item{y}{a numeric vector or a one column matrix of responses. It represents the response variable for each observation.}

\item{ncp}{a positive integer. \code{ncp} is the number of Dual-SPLS components.}

\item{ppnu}{a positive real value, in \code{[0,1]}. \code{ppnu} is the desired
proportion of variables to shrink to zero for each component (see Dual-SPLS methodology).}

\item{nu2}{a positive real value. \code{nu2} is a constraint parameter.}

\item{verbose}{a boolean value indicating whether or not to diplay the iterations steps.}
}
\value{
A \code{list} of the following attributes
\item{Xmean}{the mean vector of the predictors matrix \code{X}.}
\item{scores}{the matrix of dimension \eqn{n x ncp} where \code{n} is the number of observations.The \code{scores} represents
the observations in the new component basis computed by the compression step
of the Dual-SPLS.}
\item{loadings}{the matrix of dimension \eqn{p x ncp} that represents the Dual-SPLS components.}
\item{Bhat}{the matrix of dimension \eqn{p x ncp} that regroups the regression coefficients for each component.}
\item{intercept}{the vector of intercept values for each component.}
\item{fitted.values}{the matrix of dimension \eqn{n x ncp} that represents the predicted values of \code{y}}
\item{residuals}{the matrix of dimension \eqn{n x ncp} that represents the residuals corresponding
to the difference between the responses and the fitted values.}`
\item{lambda1}{the vector of length \eqn{ncp} collecting the parameters of sparsity used to fit the model at each iteration.}
\item{zerovar}{the vector of length \eqn{ncp} representing the number of variables shrinked to zero per component.}
}
\description{
The function \code{d.spls.lasso} performs dimensional reduction as in PLS methodology combined to variable selection via the
Dual-SPLS algorithm with the norm \eqn{\Omega(w)=\lambda_1 \|w\|_1 +\lambda_2 \|Xw\|_2 + \|w\|_2}.
}
\details{
This procedure computes latent sparse components that are used in a regression model.
The optimization problem of the Dual-SPLS regression for the norm \eqn{\Omega(w)=\lambda_1 \|w\|_1 +\lambda_2 \|Xw\|_2 + \|w\|_2}
comes from the Dual norm defintion of \eqn{\Omega(w)}. Indeed, we are searching for \code{w} that goes
with
\deqn{\Omega^*(z)=max_w(z^Tw) \text{ s.t. } \Omega(w)=1}
Noting that \eqn{\lambda_1} is an initial shrinkage parameter that imposes sparsity, the Dual-SPLS does not rely
on the value of \eqn{\lambda_1} but instead proceeds adaptively by choosing the propotion of zeros that the user
would like to impose in the coefficients. Which leads to the compuation of a secondary shrinkage parameter \eqn{\nu_1}.

The solution of this problem is
\deqn{\dfrac{w_j}{\|w\|_2}=\dfrac{1}{\mu} \delta_{N_{2,j}} (|z_{N_{2},\nu_2}|-\nu_1)_+}
Where
\itemize{
\item \eqn{\delta_{N_{2,j}}} is the vector of signs of the product between \eqn{(\nu_2 N_2^TN_2+I_p)^{-1}} and z.
\item \eqn{\mu} is a parameter that guarentees the constraint of \eqn{\Omega(w)=1}
\item \eqn{\nu_1} is the shrinkage parameter.
\item \eqn{z_{N_{2},\nu_2}} is the product between \eqn{(\nu_2 N_2^TN_2+I_p)^{-1}} and z.

}
}
\examples{
### load dual.spls library
library(dual.spls)
### parameters
n <- 200
p <- 100
nondes <- 150
sigmaondes <- 0.01
data=d.spls.simulate(n=n,p=p,nondes=nondes,sigmaondes=sigmaondes)

X <- data$X
y <- data$y


#fitting the model
mod.dspls <- d.spls.ridge(X=X,y=y,ncp=10,ppnu=0.9,nu2=0.05,verbose=TRUE)

str(mod.dspls)

### plotting the observed values VS predicted values
plot(y,mod.dspls$fitted.values[,6], xlab="Observed values", ylab="Predicted values",
 main="Observed VS Predicted for 6 components")
points(-1000:1000,-1000:1000,type='l')

### plotting the regression coefficients
par(mfrow=c(3,1))
i=6
nz=mod.dspls$zerovar[i]
plot(1:dim(X)[2],mod.dspls$Bhat[,i],type='l',
    main=paste(" Dual-SPLS (ridge), ncp =", i, " #0coef =", nz, "/", dim(X)[2]),
    ylab='',xlab='' )
inonz=which(mod.dspls$Bhat[,i]!=0)
points(inonz,mod.dspls$Bhat[inonz,i],col='red',pch=19,cex=0.5)
legend("topright", legend ="non null values", bty = "n", cex = 0.8, col = "red",pch=19)
}
\seealso{
\code{browseVignettes("dual.spls")}
}
\author{
Louna Alsouki François Wahl
}
