% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.spls.simulate.R
\name{d.spls.simulate}
\alias{d.spls.simulate}
\title{Simulation of a data}
\usage{
d.spls.simulate(n=200,p=100,nondes=50,sigmaondes=0.05,sigmay=0.5)
}
\arguments{
\item{n}{a positive integer. \code{n} is the number of observations. Default value is 200.}

\item{p}{a positive real value or a vector of length 2 representing the number of variables. Default value is 100.}

\item{nondes}{a positive integer or a vector of length 2. \code{nondes} is the number of Guassians in each mixture. Default value is 50.}

\item{sigmaondes}{a positive real value or a vector of length 2. \code{sigmaondes} is the standard deviation of the Gaussians. Default value is 0.05.}

\item{sigmay}{a real value. \code{sigmay} is the uncertainty on \code{y}. Default value is 0.5.}
}
\value{
A \code{list} of the following attributes
\item{X}{the preditors matrix. If length(p)=2, it is the concatunated predictor matrix.}
\item{y}{the response vector.}
\item{y0}{the reponse vector without noise \code{sigmay}.}
\item{sigmay}{the uncertainty on \code{y}.}
\item{sigmaondes}{the standard deviation of the Gaussians.}
}
}
\description{
The function \code{d.spls.simulate} simulates one or two mixtures of \code{nondes} Gaussians representing one or two
predictors matrices \code{X} or \code{X1} and \code{X2}
of \code{n} observations and \code{p} or \code{p1} and \code{p2} variables and a response vector \code{y}.
}
\details{
The predictors matrices \code{X} or \code{X1} and \code{X2} represent each a mixture of \code{nondes} Gaussians characterized by means and
amplitudes chosen randomly and a standard deviation set to \code{sigmaondes}.

The response vector \code{y} is the sum of 4 intervals of each predictor matrix to which we add noise of \code{sigmay}.
}
\examples{
### load dual.spls library
library(dual.spls)
####one predictors matrix
### parameters
n <- 100
p <- 50
nondes <- 20
sigmaondes <- 0.5
data1=d.spls.simulate(n=n,p=p,nondes=nondes,sigmaondes=sigmaondes)

Xa <- data1$X
ya <- data1$y

###plotting the data
plot(Xa[1,],type='l',ylim=c(0,max(Xa)),main='Data', ylab='Xa',col=1)
for (i in 2:n){
 lines(Xa[i,],col=i)
}

####two predictors matrix
### parameters
n <- 100
p <- c(50,100)
nondes <- c(20,30)
sigmaondes <- c(0.05,0.02)
data2=d.spls.simulate(n=n,p=p,nondes=nondes,sigmaondes=sigmaondes)

Xb <- data2$X
X1 <- Xb[,(1:p[1])]
X2 <- Xb[,(p[1]+1):p[2]]
yb <- data2$y

###plotting the data
plot(Xb[1,],type='l',ylim=c(0,max(Xb)),main='Data', ylab='Xb',col=1)
for (i in 2:n){
 lines(Xb[i,],col=i)
}

###plotting the data
plot(X1[1,],type='l',ylim=c(0,max(X1)),main='Data X1', ylab='X1',col=1)
for (i in 2:n){
 lines(X1[i,],col=i)
}

###plotting the data
plot(X2[1,],type='l',ylim=c(0,max(X2)),main='Data X2', ylab='X2',col=1)
for (i in 2:n){
 lines(X2[i,],col=i)
}
}
\seealso{
\code{browseVignettes("dual.spls")}
}
\author{
Louna Alsouki François Wahl
}
