% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.spls.type.R
\name{d.spls.type}
\alias{d.spls.type}
\title{Splitting the observations into groups}
\usage{
d.spls.type(y, ncells)
}
\arguments{
\item{y}{a numeric vector.}

\item{ncells}{a positive integer. \code{ncells} is the number of subsamples desired.}
}
\value{
A vecor of index specifying each observation belonging to wich group index.
}
\description{
The internal function \code{type} divides the response vector \code{y} in \code{cells}
of equal range and attributes a index type to the observations according to the corresponding cell
}
\seealso{
\code{\link[=d.spls.calval]{d.spls.calval()}}, \code{\link[=d.spls.FWLAS.calval]{d.spls.FWLAS.calval()}}, \code{browseVignettes("dual.spls")}
}
\author{
Louna Alsouki François Wahl
}
\keyword{internal}
